Split editor.js tests to editor and baseeditor.
